Enable S3 NIRISS
